requirements.txt
streamlit
langchain
pypdf
google.generativeai
python-dotenv